change datatype long to int

+++ src/main/java/com/ucc/model/Actor.java
@@ -1,14 +1,14 @@
 package com.ucc.model;
 
 public class Actor {
-    private Long actor_id;
+    private int actor_id;
     private String first_name;
     private String last_name;
     
-    public Long getActor_id() {
+    public int getActor_id() {
         return actor_id;
     }
-    public void setActor_id(Long actor_id) {
+    public void setActor_id(int actor_id) {
         this.actor_id = actor_id;
     }
     public String getFirst_name() {
@@ -23,7 +23,7 @@ public class Actor {
     public void setLast_name(String last_name) {
         this.last_name = last_name;
     }
-    
+
     @Override
     public String toString() {
         return "Actor [actor_id=" + actor_id + ", first_name=" + first_name + ", last_name=" + last_name + "]";
